Fix connection prompt not displaying any domain
